Completed Palindrom Recursion Challenge
The trick was that the line where you call the function again
needs to be a return. Otherwise it calls and seems to work, but
never returns true or false - it returns None. This was confusing
at first as I put debugging print statements everywhere and it
was recursing, but I was getting back a "None" instead of
"True". My clue something was up was that when it was "False", it worked.